leo-worker/archive-2025-12-17/root/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "app-factory"
version = "0.1.0"
description = "AI App Factory - Generate full-stack applications from prompts"
requires-python = ">=3.12"
dependencies = [
    # Extracted libraries (using local path for now)
    "cc-agent==1.10.0",
    "cc-tools==1.7.0",
    # Core dependencies
    "claude-agent-sdk>=0.1.4",
    "anyio>=4.0.0",
    "python-dotenv>=1.0.0",
    # Dependencies still needed by app-factory directly
    "fastmcp>=2.3.4",
    "psutil>=5.9.0",
    "playwright>=1.40.0",
    "aiohttp>=3.8.0",
    # Context awareness dependencies (some may be duplicated by cc-agent)
    "mcp-server-tree-sitter>=0.1.0",
    "mem0ai>=0.1.0",
    "qdrant-client>=1.11.0",
    "neo4j>=5.0.0",
    "pydantic>=2.0.0",
    "graphiti-core>=0.3.0",
    # DALL-E MCP server dependencies (may be duplicated by cc-tools)
    "openai>=1.0.0",
    "httpx>=0.24.0",
    "pillow>=10.0.0",
    # FastAPI agent server dependencies
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "requests>=2.31.0"
]

[project.scripts]
# MCP tool commands from cc-tools
mcp-integration-analyzer = "cc_tools.integration_analyzer.server:main"
mcp-cwd-reporter = "cc_tools.cwd_reporter.server:main"
mcp-shadcn = "cc_tools.shadcn.server:main"
mcp-dev-server = "cc_tools.dev_server.server:main"
mcp-browser = "cc_tools.browser.server:main"
mcp-package-manager = "cc_tools.package_manager.server:main"
mcp-build-test = "cc_tools.build_test.server:main"
# Context awareness MCP tools
mcp-mem0 = "cc_tools.mem0.server:main"
mcp-tree-sitter = "cc_tools.tree_sitter.server:main"
mcp-context-manager = "cc_tools.context_manager.server:main"
mcp-graphiti = "cc_tools.graphiti.server:main"
mcp-route-testing = "cc_tools.route_testing:main"

[tool.hatch.build.targets.wheel]
packages = ["src/app_factory", "src/app_factory_leonardo_replit"]

# Use PRIVATE CodeArtifact repository
[[tool.uv.index]]
name = "internal"
# URL is set by environment variable UV_INDEX_INTERNAL_URL
# Run: source auth-codeartifact.sh
url = "https://cc-internal-855235011337.d.codeartifact.us-east-1.amazonaws.com/pypi/python-internal/simple/"
explicit = true

[tool.uv.sources]
# Use local submodules for development (fast iteration, no AWS auth needed)
# For CI/production, override with: UV_OVERRIDE_cc_agent=... UV_OVERRIDE_cc_tools=...
cc-agent = { path = "vendor/cc-agent", editable = true }
cc-tools = { path = "vendor/cc-tools", editable = true }
